Clean up and club repository abstract
Made the Club Repository abstract, and also organized the files a little bit
