+++ convene/lib/screens/login/state/login_controller.dart
@@ -1,21 +1,20 @@
 import 'package:authentication_repository/authentication_repository.dart';
+import 'package:convene/services/authentication/models/models.dart';
 import 'package:formz/formz.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:state_notifier/state_notifier.dart';
 
-import '../../authentication/models/models.dart';
-import 'sign_up_state.dart';
+import 'login_state.dart';
 
-class SignUpController extends StateNotifier<SignUpState> {
-  SignUpController(this.read, [SignUpState state])
-      : super(state ?? const SignUpState());
+class LoginController extends StateNotifier<LoginState> {
+  LoginController(this.read, [LoginState state])
+      : super(state ?? const LoginState());
 
   /// The `ref.read` function
   final Reader read;
 
   void emailChanged(String value) {
     final email = Email.dirty(value);
-
     state = state.copyWith(
       email: email,
       status: Formz.validate([email, state.password]),
@@ -30,14 +29,14 @@ class SignUpController extends StateNotifier<SignUpState> {
     );
   }
 
-  Future<void> signUpFormSubmitted() async {
+  Future<void> loginWithCredentials() async {
     state = state.copyWith(hasSubmitted: true);
 
     if (!state.status.isValidated) return;
 
     state = state.copyWith(status: FormzStatus.submissionInProgress);
     try {
-      await read(authRepositoryProvider).signUp(
+      await read(authRepositoryProvider).logInWithEmailAndPassword(
         email: state.email.value,
         password: state.password.value,
       );
