+++ convene/lib/screens/login/state/login_state.dart
@@ -1,16 +1,15 @@
+import 'package:convene/services/authentication/authentication.dart';
 import 'package:formz/formz.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../../authentication/models/models.dart';
-
-part 'sign_up_state.freezed.dart';
+part 'login_state.freezed.dart';
 
 @freezed
-abstract class SignUpState with _$SignUpState {
-  const factory SignUpState({
+abstract class LoginState with _$LoginState {
+  const factory LoginState({
     @Default(Email.pure()) Email email,
     @Default(Password.pure()) Password password,
     FormzStatus status,
     @Default(false) bool hasSubmitted,
-  }) = _SignUpState;
+  }) = _LoginState;
 }
