+++ convene/packages/club_repository/lib/club_repository.dart
@@ -1,4 +1,4 @@
 library club_repository;
 
-export 'src/club_repository.dart';
 export 'providers.dart';
+export 'src/club_repository.dart';
