+++ convene/packages/club_repository/lib/providers.dart
@@ -1,3 +1,4 @@
+import 'package:club_repository/src/firestore_club.dart';
 import 'package:riverpod/riverpod.dart';
 
 import 'club_repository.dart';
@@ -5,5 +6,5 @@ import 'club_repository.dart';
 /// Provides methods to read and write to the User database. Uses the
 /// [UserDatabase] object
 final clubRepositoryProvider = Provider<ClubRepository>((ref) {
-  return ClubRepository(ref.read);
+  return FirestoreClub(ref.read);
 });
